All code now should work, just run main.py
